# Conflicts:
#	crm/src/main/java/com/hy/crm/controller/BusinessController.java
#	crm/src/main/java/com/hy/crm/controller/UserController.java
#	crm/src/main/java/com/hy/crm/entity/Business.java
#	crm/src/main/java/com/hy/crm/entity/Clien.java
#	crm/src/main/java/com/hy/crm/entity/User.java
#	crm/src/main/java/com/hy/crm/mapper/UserMapper.java
#	crm/src/main/java/com/hy/crm/mapperSql/StudentSql.java
#	crm/src/main/java/com/hy/crm/service/impl/BusinessServiceImpl.java
#	crm/src/main/java/com/hy/crm/service/impl/UserServiceImpl.java
#	crm/src/main/webapp/1.html
#	crm/src/main/webapp/business/business.html
#	crm/src/main/webapp/business/businessadd.html
#	crm/src/main/webapp/business/businessupdate.html
#	crm/src/main/webapp/business/wobusiness.html

+++ crm/src/main/java/com/hy/crm/controller/After_saleController.java
@@ -57,7 +57,7 @@ public class After_saleController {
      */
     @RequestMapping("/statistics.do")
     @ResponseBody
-    public List<Long> statistics(){
+    public List<Integer> statistics(){
         return after_saleService.counts();
     }
